Implemented database environment variables correctly. Renamed dbConfig file.
