+++ src/api/config/dbConfig.js
@@ -1,9 +1,7 @@
-import {env_var} from 'env_var'
-
 const Sequelize = require('sequelize');
-const db = new Sequelize(env_var.DB_NAME, env_var.DB_USER, env_var.DB_PASSWORD, {
-  host: env_var.DB_HOST,
-  dialect: env_var.DB_DIALECT,
+const db = new Sequelize(process.env.DB_NAME, process.env.DB_USER, process.env.DB_PASSWORD, {
+  host: process.env.DB_HOST,
+  dialect: process.env.DB_DIALECT,
   logging: false,
 
   pool: {
